Check if the block instance belongs to the view Bug#1233500
When editing a page, check the ownership of blocks before
1. update the configuration data
2. delete/move them

+++ htdocs/lib/view.php
@@ -2062,6 +2062,10 @@ class View {
         }
         require_once(get_config('docroot') . 'blocktype/lib.php');
         $bi = new BlockInstance($values['id']); // get it so we can reshuffle stuff
+        // Check if the block_instance belongs to this view
+        if ($bi->get('view') != $this->get('id')) {
+            throw new AccessDeniedException(get_string('blocknotinview', 'view', $bi->get('id')));
+        }
         db_begin();
         $bi->delete();
         $this->shuffle_cell($bi->get('row'), $bi->get('column'), null, $bi->get('order'));
@@ -2087,6 +2091,10 @@ class View {
         }
         require_once(get_config('docroot') . 'blocktype/lib.php');
         $bi = new BlockInstance($values['id']);
+        // Check if the block_instance belongs to this view
+        if ($bi->get('view') != $this->get('id')) {
+            throw new AccessDeniedException(get_string('blocknotinview', 'view', $bi->get('id')));
+        }
         db_begin();
         // moving within the same column and row
         if ($bi->get('row') == $values['row'] && $bi->get('column') == $values['column']) {
@@ -2255,6 +2263,10 @@ class View {
     public function configureblockinstance($values) {
         require_once(get_config('docroot') . 'blocktype/lib.php');
         $bi = new BlockInstance($values['id']);
+        // Check if the block_instance belongs to this view
+        if ($bi->get('view') != $this->get('id')) {
+            throw new AccessDeniedException(get_string('blocknotinview', 'view', $bi->get('id')));
+        }
         return $bi->render_editing(true);
     }
